requirements.txt
flask
tensorflow
pillow
numpy
gunicorn 
